+++ src/main/java/com/nucleus/customer/model/Customer.java
@@ -1,13 +1,7 @@
 package com.nucleus.customer.model;
 
 import com.nucleus.loanapplications.model.LoanApplications;
-import org.hibernate.validator.constraints.NotEmpty;
-import org.springframework.format.annotation.DateTimeFormat;
-
 import javax.persistence.*;
-import javax.validation.constraints.Min;
-import javax.validation.constraints.NotNull;
-import javax.validation.constraints.Pattern;
 import java.time.LocalDate;
 import java.util.List;
 import java.util.Set;
@@ -32,33 +26,24 @@ public class Customer {
     private String customerCode;
 
     @Column(name="first_name",length = 30,nullable = false)
-    @Pattern(regexp = "^[A-Za-z]{3,30}$", message = "Please enter a valid First Name")
     private String firstName;
 
     @Column(name="last_name",length = 30,nullable = false)
-    @Pattern(regexp = "^[A-Za-z]{3,30}$", message = "Please enter a valid Last Name")
     private String lastName;
 
     @Column(name="date_of_birth" , nullable = false)
-    @DateTimeFormat(pattern="dd-MM-yyyy")
-    @NotNull(message = "Please Enter A Date")
     private LocalDate dateOfBirth;
 
     @Column(name="nationality",length = 30, nullable = false)
-    @NotNull(message = "Please select a valid Nationality")
     private String nationality;
 
     @Column(name="occupation_type",length = 30, nullable = false)
-    @NotNull(message = "Please select a valid Occupation type")
     private String occupationType;
 
     @Column(name="total_work_experience",length = 2)
-    @NotNull(message = "Please enter a positive number")
-    @Min(value=0)
     private Integer totalWorkExperience;
 
     @Column(name="organization_name",length = 30)
-    @NotEmpty(message = "Organization name cannot be empty")
     private String organizationName;
 
     @OneToMany(fetch = FetchType.LAZY , mappedBy="customerCode")
