+++ src/main/java/com/nucleus/customerservice/repaymentschedulereport/service/RepaymentScheduleReportServiceImpl.java
@@ -13,10 +13,6 @@ import java.time.LocalDate;
 import java.util.ArrayList;
 import java.util.List;
 
-/**
- * This class acts as a Service layer for getting Repayment Schedule Report .
- *
- */
 @Service
 public class RepaymentScheduleReportServiceImpl implements RepaymentScheduleReportService {
 
@@ -33,12 +29,11 @@ public class RepaymentScheduleReportServiceImpl implements RepaymentScheduleRepo
         transactionTemplate = new TransactionTemplate(transactionManager);
     }
 
+    static LocalDate returnDate(String date) {
+        LocalDate dt = LocalDate.parse(date);
+        return dt;
+    }
 
-    /**
-     * Retrieves a Repayment Schedule Report by loan application number from the database.
-     * @param loanApplicationNumber is the Loan Application Number of the Repayment Schedule to be retrieved.
-     * @return List of given Repayment Schedule if exists, else null.
-     */
     @Override
     public List<RepaymentSchedule> getRepaymentScheduleReport(int loanApplicationNumber) {
         List<RepaymentSchedule> repaymentScheduleReport = repaymentScheduleDAO.getRepaymentScheduleReport(loanApplicationNumber);
