+++ src/main/java/com/nucleus/repaymentschedule/dao/RepaymentScheduleDAOImpl.java
@@ -12,10 +12,6 @@ import org.springframework.transaction.annotation.Transactional;
 import javax.persistence.Query;
 import java.util.List;
 
-/**
- * This class acts as a Database layer for RepaymentSchedule related operations.
- *
- */
 @Repository
 @Transactional
 public class RepaymentScheduleDAOImpl implements RepaymentScheduleDAO
@@ -24,11 +20,6 @@ public class RepaymentScheduleDAOImpl implements RepaymentScheduleDAO
     @Autowired
     private SessionFactory sessionFactory;
 
-
-    /**
-     * Adds a Repayment Schedule to the database.
-     * @param repaymentSchedule is the RepaymentSchedule to be saved.
-     */
     @Override
     public int addRepaymentSchedule(RepaymentSchedule repaymentSchedule) {
