+++ src/main/webapp/navbar.jsp
@@ -101,9 +101,6 @@
                         <a class="dropdown-item" href="<%= request.getContextPath()%>/receiptBOD/run">Receipt BOD</a>
                         <a class="dropdown-item" href="<%= request.getContextPath()%>/main/loanClosureBod">Loan Closure BOD</a>
 
-                        <a class="dropdown-item" href="<%= request.getContextPath()%>/addRepaymentReport">Temp Add Loan</a>
-                        <a class="dropdown-item" href="<%= request.getContextPath()%>/main/loanClosureForm">Loan Closure Search</a>
-
                     </div>
                 </li>
                 <li class="nav-item mx-2">
