--- src/test/java/LoanApplicationTest.java
@@ -1,49 +1,0 @@
-import com.nucleus.config.TestConfig;
-import com.nucleus.customer.model.Address;
-import com.nucleus.customer.model.Customer;
-import com.nucleus.loanapplications.model.LoanApplications;
-import com.nucleus.loanapplications.service.LoanApplicationService;
-import com.nucleus.loanapplications.service.NewLoanApplicationService;
-import org.junit.Test;
-import org.junit.runner.RunWith;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.test.context.ContextConfiguration;
-import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;
-import org.springframework.test.context.web.WebAppConfiguration;
-
-import javax.enterprise.inject.New;
-
-import java.time.LocalDate;
-
-import static org.junit.Assert.*;
-
-@RunWith(SpringJUnit4ClassRunner.class)
-@WebAppConfiguration
-@ContextConfiguration(classes = TestConfig.class)
-
-public class LoanApplicationTest {
-    @Autowired
-    LoanApplicationService loanApplicationService;
-
-    @Test
-    public void getAllLoanApplicationsListTest() {
-        assertNull(loanApplicationService.getAllLoanApplicationsList());
-    }
-
-    @Test
-    public void getLoanApplicationIdTest() {
-        assertNull(loanApplicationService.getLoanApplicationId(1));
-    }
-
-    @Test
-    public void deleteLoanApplicationIdTest() {
-        assertTrue(loanApplicationService.deleteLoanApplicationId(123456));
-    }
-
-    @Test
-    public void updateLoanApplicationTest(){
-        assertNull(loanApplicationService.updateLoanApplication(new LoanApplications()));
-    }
-
-
-}
